PathEnumerator impl: Workspace implements Enumerable
This implements a new class that provides an enumerator over a
`PathTree`, and allows `Workspace` to include & implement `Enumerable`.

The new class includes logic to prevent infinite loops from symlink
directories that could point to their own ancestors.

The behavior of the enumerator is that every node in the `PathTree`
underlying the `Workspace` is an item in the collection: i.e. a block
passed to `#each` will receive both directories & the files within them.

Items in the enumerator are `Pathname` instances