Init: respect excludes when enabling engines
`ConfigGenerator` now uses the new `Workspace` enumeration when looking
for files that match engine enable patterns. This fixes one small bug &
comes with a performance improvement as a side benefit

The bug fix is that because we used to walk the *entire* tree for files
matching an engine's patterns, you could enable an engine whose only
matching files were in auto-exclude paths. e.g. a project under
`vendor/` might have some JavaScript files used for local scripting or
something while your own code had no JS at all.

The performance benefit comes as the result of the lazily-evaluated
nature of the `Workspace` enumeration. Previously we were shelling out
to `find` to get a list of all files in the tree, which was usually not
noticeable but on certain cases could cause a noticeable lag running
`init`.