Init: respect excludes when enabling engines
`ConfigGenerator` now uses the new `Workspace#include?` when looking
for files that match engine enable patterns. This fixes one small bug &
comes with a performance improvement as a side benefit.

The bug fix is that because we used to check the *entire* tree for files
matching an engine's patterns, you could enable an engine whose only
matching files were in auto-exclude paths. E.g. a project under
`vendor/` might have some JavaScript files used for local scripting or
something while your own code had no JS at all.

The performance benefit comes for similar reasons to why the recent
refactor of workspace calculation with exclude paths was an improvement:
we touch less data && less of the filesystem. The previous
implementation shelled out to `find` to get a list of every file in
tree, and then matched the engine's `enable_regexps` against that large
array. The new algorithm globs for just files matching the engine's
`enable_patterns`, and then walks the `Workspace`'s `PathTree` to see if
that file should considered to be in the workspace.

In my local test cases, peformance of `init` overall is ~30% better than
it was before.