Workspace#include?: test whether a path is in a workspace
Useful for testing presence of a file within a workspace. This will be
used by `ConfigGenerator` for testing which engines should be enabled.

This is just a pass-through to PathTree#includes?, also added here: the
method walks a PathTree (without changing its internal path collection)
to tell you if a given path exists within the tree.