ConfigGenerator: raise error if workspace is empty
We were just returning an empty set of paths to enable engines on in
init when everything was excluded and/or the dir was empty, but this
strikes me as incorrect. If there's really nothing for us to look at, it
seems more reasonable to call that an error condition & exit
immediately.