Allow excluding everything from a workspace
Previously, the way that `PathTree` and its nodes detected whether to
return their child paths or not was a bit surprising in that it
effectively "reset" itself if you removed *everything*. So if you had an
`exclude_path` of "*", as a fairly dumb example, engines would actually
get an `include_paths` set of `["./"]`.

Although certainly an edge case, this is still surprising & undesirable
behavior. This changes the `DirNode'`s internal state tracking so that
it remembers whether it *ever* had children, instead of whether it
*currently* does. This allows excluding everything and getting a correct
empty path set as a result.

Arguably this "feature" isn't very useful without also telling the user
something is wrong, so I'm going to change that next.